VNC-104 Move frame rate helper under core utilities
